urmat

+++ jdbc/jdbcAndHibernate/src/main/java/peaksoft/dao/UserDao.java
@@ -10,7 +10,7 @@ public interface UserDao {
 
     void dropUsersTable();
 
-    void saveUser(String name, String lastName, byte age);
+    void saveUser(String name, String lastName, int age);
 
     void removeUserById(long id);
