+++ jdbc/jdbcAndHibernate/src/main/java/peaksoft/service/UserService.java
@@ -10,11 +10,12 @@ public interface UserService {
 
     void dropUsersTable();
 
-    void saveUser(String name, String lastName, byte age);
+    void saveUser(String name, String lastName, int age);
 
     void removeUserById(long id);
 
     List<User> getAllUsers();
+
 
     void cleanUsersTable();
 }
